Added some auth to UserController.

+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UserRequest;
 use App\Models\User; 
 use App\Providers\UserProvider;
+use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
@@ -13,15 +14,19 @@ class UserController extends Controller
     ) {}
 
     // Display a listing of the resource.
-    public function index()
+    public function index(Request $request)
     {
+        $this->authorize('index', User::class);
+
         $users = User::all();
         return view('users.index', compact('users'));
     }
 
     // Show the form for creating a new resource.
-    public function create()
+    public function create(Request $request)
     {
+        $this->authorize('create', User::class);
+
         $user = new User();
 
         return view('users.create', $this->userProvider->populateModel($user));
@@ -30,6 +35,8 @@ class UserController extends Controller
     // Store a newly created resource in storage.
     public function store(UserRequest $request)
     {
+        $this->authorize('store', User::class);
+
         $user = new User();
     
         $this->userProvider->persist($request, $user);
@@ -45,14 +52,18 @@ class UserController extends Controller
     }
 
     // Show the form for editing the specified resource.
-    public function edit(User $user)
+    public function edit(Request $request, User $user)
     {
+        $this->authorize('edit', $user);
+
         return view('users.edit', $this->userProvider->populateModel($user));
     }
 
     // Update the specified resource in storage.
     public function update(UserRequest $request, User $user)
     {
+        $this->authorize('update', $user);
+
         $this->userProvider->persist($request, $user);
 
         return redirect()->route('users.index')
@@ -60,8 +71,10 @@ class UserController extends Controller
     }
 
     // Remove the specified resource from storage.
-    public function destroy(User $user)
+    public function destroy(Request $request, User $user)
     {
+        $this->authorize('destroy', $user);
+        
         $user->delete();
         return redirect()->route('users.index')
             ->with('success', 'User deleted successfully.');
